// RasProyecto/src/app/core/models/usuario.model.ts
export interface Usuario {
  id: string;
  nombre: string;
  correo: string;
  contrasena: string;
  programa: string;
  telefono: string;
  sede: string;
  cedula: string;
  miembroDesde: string;
  etapa: string;
  iniciales: string;
  colorAvatar: string;
  rating: number;
  totalResenas: number;
  totalTrueques: number;
  totalArticulos: number;
}